enhance: move the MRB1 segcore pin out of pkg, and drop the blob digests
Review follow-up on this branch. Four things, no behaviour change.

pkg/v3 is a published standalone module, and its test suite had grown a
dependency on the segcore source tree: the parity test walked up for
internal/core/CMakeLists.txt and read RoaringMembership.{h,cpp}. The only
thing holding it there was the seven unexported kPortable* constants. Those
are RoaringFormatSpec values rather than a Milvus contract -- two of them are
derived, not chosen -- and each side is already held to them behaviourally at
the boundary: the bitmap-container threshold by the fixtures each side
AcceptsValidPortableContainerEncodings at three and four containers, the
minimum-entry bound by RejectsUnsupportedRoaring32CookieAfterPrefixBound.
That is the stronger check anyway, since a text comparison cannot see a use
site. They stay in the completeness list so a constant added to the .cpp
still has to be classified.

The nine MRB1 envelope and admission constants are all exported, so the pin
moves to internal/parser/planparserv2 as TestRoaringSegcoreConstantsMatch. It
already owns the Go-to-Go half, and it is in segcore's own module, so reading
those sources crosses no boundary. Not finding them is now a failure rather
than a skip, for the same reason.

mentionedInSegcore matched a substring, so renaming kMaxBodySize to
kMaxBodySizeBytes reported "widen cppIntConstant" rather than "no longer
declares kMaxBodySize" -- the same conflation an earlier round fixed for a
different input. Whole-word now, verified by injecting both renames. The
30-entry width table is also gone: an unlisted declared type is an error
instead of an assumed 64 bits, which is what the table was guessing at.

containerEncodingsBlobSHA was one digest literal copied into three modules,
and by its own comment it could not prove the three member sets were equal --
updating one copy and its digest left the others passing. All three assert the
member count instead, which is what separates 4097 from 4096 when the
container counts and the body length do not. pkg/util/roaringfilter, which
classifies the containers directly, keeps that as the real check.

Also: the pkg tests still named themselves TestParse* for a package that no
longer exports Parse, and the skip-index TODO in RoaringFilterExpr.cpp is
trimmed to the issue reference and the two's-complement caveat.
